Need to try the same for card buttons. I need to think about how my cards will interact with the trident they belong to and the other cards in the same trident. CardButtons will be used only in the trident building phase i think, once i’ve assigned them to the trident and moved to the game stage, i should use triButtons instead of cardButtons. So the TriButton should be much more heavily used that the CardButtons, cardButtons will need logic related to being added to the trident, the empty trident hand will also have to use cardButtons, so cardButtons must be able to be used without a card value.
Card will have orientation, same as trident, but it will be the implicit orientation of the resulting trident, it will also have a position which will refer to the left right or vertical (top or bottom) position. The actual x, y position of the card should be taken as being the bottom left of the imaginary trident it would be in,

I’ve made methods do detect if the touch lands in the card, for each orientation and position. Seems to work well,