Setting up the touchLogic of the triButton and cardButton classes to use a switch statement to identify which stage the button is on and then call that stage’s touchLogic method passing in the button’s index value so we know which button was pressed. Now i need a switch statement in the stage’s touchLogic method to do different action depending on which button index was passed in
Instead of my method of incrementing an int when creating buttons and using that to get the button from the array of buttons, i’ve now used enums instead, which is better as they are incremented automatically, can use the .ordinal to get the value of the enum, which will be 0 then 1  then 2 etc for each successive one
However i’ve read on the internet it’s not good practice to use ordinal in this way, partly because java could change and ordinals could be changed... also if i don’t declare the enums in the same order i create the buttons then when i try and get the button from the array using that enum it might get the wrong one, to help solve this i could try adding the button to the array at that enum index position, if they are out of order then it will cause an error, which i will be able to fix, alternatively could i use a hash map?
I’ve now learned that you can declare an enum to have a value so instead of EXIT, i can do EXIT(0), then i don’t have to use the ordinal,so the value will be identical to the ordinal because i have chosen to assign 0 rather than for example 1, but it’s better practice to use this value rather than ordinal, you can even add a constructor for the enum to have this value increment automatically.
Tested this now, if you insert the buttons in the wrong order then the game crashes which is what i wanted.

Ok so now we have two types of buttons, we can click the buttons, even if the bounding boxes are overlapping, the TitleStage has the touch method for it’s buttons set out, can copy that to the other stages, can have individual methods for each button, so now i think i can flesh out all the stages.