In terms of fleshing out the stages, title screen just needs a few buttons, Options screen will need loads of buttons, but there’s no point in making them all yet, additionally i need a method to automatically adjust the layout when adding more buttons. Matchmaking stage, this is probably something which i’ll leave for now, so the deal is what i should work on next, The deal stage will be arrived at after matchmaking, it’s the first part of the actual game, although it’s mainly just an animation i suppose. Check out the design doc for more info on deal stage. basically , first i need a deal button, click this, then i need to display all the cards on screen spread out, then flip them over to so you can’t see the value (if that option is enabled) make them all move to the same location, then move them all into the card array of each player, so display 2 or 3 card arrays So i need a card class, and a way to shuffle them, a simple class, with all static members and methods, called deck, has an array of 52 card classes, a card should have an index and a value, then we just need an array of 52 unique random integers between 0 and 51 which will refer to the cards in the deck
Each player will have a hand of cards, but i should cater for if the other player leaves, so i should store a local copy of all player data on each person’s device, maybe then we can resume the session without losing progress
Need to resume from the setUpCards method in tridentBuildingStage and associated methods, the trident stage will set up all 52 cardButtons to take input, but i need to set the position of each button based on which player it was dealt to, also i’ll eventually need methods to figure out what cardButton is adjacent to a cardButton ,
It’s starting to get confusing, but i haven’t done so much that it’s too late to remove these bits and try again

Had a bit of a rethink about what i did yesterday, i now want to get rid of the card class, create 52 cardButtons, display these in deal stage, but they actually should belong to trident building stage, where you can interact with them too. So i need to create them and give each one the unique enum, but then after creation i’ll need to assign the card value and update position etc
I’ve set up a rectangle in which the cardbuttons will appear for the dealStage
Need a method to spread the cardbuttons across this rectangle, also consider that the rectangle can change size, if i feel like changing it later, might be best to simply assume it’s square. The cards will be arranged so that 3 cards make a trident, so there will be 17 tridents and one extra card left over, tridents are 0.86 tall as they are wide, a rectangle 2 tridents tall and 1 trident wide approximates 1280 by 720quite well actually.  But 2 tridents wide is not twice as wide as one trident , it’s only 1.5 times wider, tried a few patterns, 1280 /720 is 1.777777778 and 5 tridents tall by 4 across would be 1.732050808, so this should be fine
So arrange the cards in tridents like so
▲▼▲▼
▼▲▼▲
▲▼▲▼
 ▼’
4 tridents of 3 cards on the first 4 rows, one trident and 1 lone card on the last row
From all this we can work out how large to display the cards, for this part of the stage only.
The height of the defined rectangle / 5 /0.86 will equal the edgeLength,
Or the width of the defined rectangle /2.5 will be the edge length,
We need to use whichever is smaller

Ok so i’ve created the cardButtons in the tridentStage, gave them a position and orientation etc relevant for the dealStage specifically, drawn them on the dealStage, now i need to animate them to look like they are being dealt, before i do that i’ll have to set up the correct number of players etc
Need to find the best time to actually deal the cards to the players, make sure the random cards can’t be seen by other players
Need something similar to the CardButton.overlapAnimation
But for moving all the cards to their position in the player card hand

Need to flesh out the method amendCardsForTridentStage