Make sure the cards only get shuffled after the overlap stage, so the shuffled cards are not visible Make it so that if multiple players connect we create a player class for each of them, also create a local copy for each, Need to set the tridentbuildingstage positions for the cards, which needs to know the correct number of players, then we can animate the cards into those positions in the deal stage, The card buttons should always appear in a set order, but the value attributed to them will be random - from the deal
before the deal stage moves on to the tridentbuildingstage and before the player can see his cards  we need to resolve ‘par’ which might involve swapping cards amongst players, i’ve worked out a way to do this previously in the design doc, if i animate the cards swapping over then i need to make sure the positions of the card buttons have not actually swapped over, i can animate them moving to each other’s location, but then when they get there i can just reset them to their original position and swap the values instead

Instead of adding the cardbuttons to the player, instead add the player index to the cardbutton, then the cardbutton touchlogic can determine what happens when the player touches it
Similarly the tridentButtons can be added to the gamestage with the associated player, this means nothing will be lost on player disconnect and i won’t need a local copy of the player

When the cardbuttons move to their player’s cardhand position, i think i should emphasise the current player’s cards, they should appear larger, also need to leave room for the player trident hand, but not the opponent’s. I need to maximise the trident hand size and the card hand size for ease of use

i’ve already defined a rectangle to display the deal in, so within that constraint i need to work out how big i can get the cards and still fit them all on screen, with a trident hand that’s just as big and space for pre and post game cards, i can make the opponents card hand appear smaller as we won’t be tapping that
I have also thought about sorting the cards by suit in the past.
When the cards were spread out initially in the start of the deal stage, there were 5 rows of 4 tridents, now i need 1 row for the trident hand, 4 rows for the card hands (one row per suit)
And then extra space to display all opponent’s cards - albeit at a smaller size ... actually no, the open game mode should allow you to see the opponent’s tridents after they have been made, i don’t think it’s a good idea to show the opponent’s card hand , so 4 rows for the suits of cards you have, then 1 row for the trident hand, also don’t forget about the mirrors for the trident, making it 3 times as tall. Furthermore consider the trident hand will be long since the card hands have been split into 4 -

In detail
Trident hand - in 3 player game with no pre game cards this will have 5 tridents,
In 2 player game with pre and post game cards, this will be 8 tridents + 2 cards, which will take up at least 1 trident of space in width, assuming we put the trident hand on one row we need an extra 2 rows for the mirrors
The card hand will be split into 4 rows one for each suit, so most cards one row could possibly have is 13, which would be 4 and a third tridents
So trident hand will be 9 tridents max
Card hand will be 5 trident wide max
3 rows + 4 rows + 6 row margins

So the edge length of the tridents in the trident hand will be defined as rectangle width /5

I’ve changed the deal methods so that, say if each player has 15 cards, there will be 15 random cards which are then put in order, then added to the array, then repeat for the next player, this means when i set the value of the cardbuttons for each player, the value will be in order

I think this has worked ok, but the camera has been weird for a while, i want it to be 720 by 1280, but altering it messes with the hit detection
Think i’ve fixed the camera now, however my monitor can’t fit it all on, might be an idea to go through everything and make any measurements a percentage of the screen rather than a fixed size

Need to add in the par logic, i need to pause the animation here and add a par button