So i’m setting up this static PH number which is screen height/1280, if i had somewhere something like edgeLength = 50, i’ll change that to 50*PH and hopefully this should scale to fit the screen, i’m going to need to force the aspect ratio though, ... scrap all that, this is what i needed
spritebatch.setProjectionMatrix(camera.combined);
shaperenderer.setProjectionMatrix(camera.combined);

The desktop app resembles the phone app now , also i’m using Gdx.graphics.setWindowedMode(315, 560);
So that the desktop app actually fits on my small tv
The begin deal trident appears in a much different location
This is because it uses Gdx.graphics.getHeight, instead i should use my defined height and width of 1280 and 720, i should go through and remove all usage of this get height method
Did that now touch doesn’t work... drew the mouse x and y on screen and it seems to be inverted, thats because setToOrtho first argument was false , changed it to true, put that literally turned things upside down including text, after more testing this wasn’ the problem, set that back to false, another problem is the mouse position is not one to one with the actual mouse position, we can’t reach the far corner of the screen

Set Gdx.graphics.setWindowedMode(720/2, 1280/2);
This needs to match the viewport exactly, so if i need to reduce the window size to fit the tv screen, then i need to change the viewport to match.

That fixes the problem on desktop, but i still get the problem on android, the full screen is not being used, because my phone is more than 1280 by 720 , it’s 1080 by 1920

If i just stop using setWindowedMode, the phone will have it’s high resolution and desktop gets some small default resolution, not sure where this default is set but i guess i could change it only for desktop mode... ok i figured that out it’s in
The desktopLauncher just add
config.width=720/2;
config.height=1280/2;

So the desktop is small res, fits on screen, android is high res and it fits on it’s screen too, but things are proportionally smaller on android

Can change the desktop width and height to mimic a real world device size, but regardless of that size, the viewport should always be 16:9
If the device is not exactly the same at the viewport then i need to

This has been very confusing but now with a fitViewport i’m happy with it, desktop window size can be whatever, but if we set the viewport to 1920 by 1080, the fit viewport will sort of zoom out so that everything drawn within the 1920 by 1080 viewport can be seen on that screen. If the real world screen size is larger than the viewport that is fine too, a lot of the confusion was due to me drawing some things in relation to the Gdx.graphics.width so that when i was testing things it wasn’t working as expected.
The viewport can be anything, can be 100 by 100, 16 by 9, 1280 by 720 or 1920 by 1080, the only thing that matters is that i’m  drawing within that space, so i can set it as 1280 by 720, but i need to remember how big the screen is, don’t need to use Gdx.graphics.getWidth, just use 720,

So i do need to go through all my code and adjust the size of everything

I’ve also noticed that even if i try and set the desktop launcher size to 1280 by 720, if one of those number exceeds my actual tv size, it will be capped, which is helpful - height is capped at 639 , which means i can see all the content but the resolution is not great, so i could make things bigger so it’s easier to read, but remember i’m not actually targeting this resolution

Set immersive mode in the android launcher

Kinda wasted a lot of time on this,

Next time go through all code and resize everything, first of all the font needs to be at least twice as big, make sure nothing references Gdx.graphics.getHeight,

After that resume the deal stage, the deal animation is not right yet