I’m going to resize everything today, font size might present a bit of a problem. It’s not really feasible to allow increasing the font size in game, things would have to change size to accommodate it, also there will not be one universal font size that will be viewable on all devices, obviously the bigger the better but resolution will matter a great deal
My game will be 720 wide by 1280 tall, although these numbers are abstract, i could just as easily say 9 wide by 16 tall.

I’ve made methods to resize tridents to fit various things, but generally speaking, in gameplay the game board is 4 tridents wide, so that trident must be a bit less than 1/4 of  720, the trident array will at most 8 tridents + 2 cards, the 2 cards will be the pre and post game cards and they will be 1 trident wide be 9 tridents wide max, a row of tridents can tessellate together though so the actual overall width of several tridents is given by 0.5 edgeLength + 0.5 * edgelength * numberoftridents, which for 9 tridents will be 5.  so trident length for that should be less than 1/5 of 720 and i think that will be the minimum size of a trident length. Where possible tridents should be increased in size to allow for easier touch controls. But slightly less than 720/5 should be fine. Let’s say trident edgeLength should be 130, so if there are 5 fitting together that leaves 70 space, so 35 space either side. altitude will be about 112.6.
The most cramped screen will be the trident building screen, we need
Margin
Trident array mirror
Trident array actual
Trident array mirror2
Margin 2
Card array suit 1
cardmargin1
Card array suit 2
cardmargin2
Card array suit 3
cardmargin3
Card array suit 4
Margin 3

Thats 7 rows of tridents with height 112.6 which is 788.1, 491.9 space remaining there are 3 big margins and 3 small margins, so 110 for big margin, 54 for small margins

Now move onto the deal stage once again

The deal methods check to see which player is which, so you need to run the server or else it will crash.
Arranged the cards neatly on screen
Still not perfect, and now i have a weird error when trying to debug as if tridentbuildingstage is in the jar as well as my project .... i’ll have to look into that ...
Error went away after a restart

Finally neatened up the tridenthand and the card hands, need a method to figure out which card hand is shortest, and therefore has the most excess space to the right of it, here i will place some buttons to help build tridents, but that’s for the trident stage. After the deal is complete move seamlessly to the trident stage, the only different visually will be these new buttons for manipulation the trident hand, other not immediately noticeable differences will be the card buttons will not allow touch

Half way through the deal, i need to stop the deal, preferably when the cards overlap, now display a button called par, we can alter par up and down to alter the deal as discussed in the design doc. I need to implement that