Need to fix this error "LWJGL Application" java.util.ConcurrentModificationException The dealStage draws the player’s in MyServer, if a new player is added then we are modifying that array of players that are being drawn, this is not allowed. This error only occurs sometimes, i guess because it would have to be timed right, to fix it i could do a try and catch in the draw method..? but instead of that do this: instead of adding the other players to the friendlyList, add them to an addingList then when we are sure we are not in the update method where we iterate over the items, then we add them to the list Hopefully that works, i’m sure i’ll find out if it doesn’t work eventually
Regarding the problem with some cards displaying at 0,0, think i fixed that now

Spotted a new problem, the 3rd player, has been dealt several cards all in the first suit, the first 3 cards should be drawn on the same spot with different orientations, the orientations are all correct, but only 2 cards are drawn in the first x position, then the next card is drawn to the right ... actually this happens for player 2 as well and it seems it only did happen for player one due to the next problem below, where it was erroneously dealt card value 0

Still have the problem where the 0 value card is dealt twice, it can either be to 2 different people or the same person, this is just because cardButton playerIndex and value  default to 0 and when i emit the deal i only emit the cards that were dealt, so i was only emitting 51 cards, the for loops in the server assumed 52 cards instead of OptionsStage.numberOfPlayers*OptionsStage.cardsEach
So any cards over that amount would count as 0 value player index 0

Fixed all these problems.

New problem: if anyone clicks begin deal before all players have joined, the cards will not be dealt, fix this by disabling the button until all players are present

Now i want the deal stage to automatically progress to the trident building stage at the end of the deal

Forgot to actually do par, so once player 1 deals the cards, it needs to amend each deal considering par, then emit the deal

So complete the deal, calculate the total value of cards for each player, we know par for each player, but to get the true par we need to add the average total value of a hand,
364 is the total value of all cards
There are 52 cards so on average a card is worth 7
So to work out the true par it’s 7*cardsEach + par

For each player we should work out total value of cards they have been dealt, then work out their truePar, then work out the difference between these values. Negative difference will mean their total needs to be raised, positive difference means their total is too high and must be lowered, the only way to raise or lower the total is to swap cards with other players or with undealt cards, so i need to make a foolproof way to do this, and i would prefer to swap as few cards as possible.