The auto build and wild card buttons seem to work, i think there is just one more thing i need to work out regarding this, if you set the wild cad suit, say suit 2, then autobuild a trident, of a specific suit say suit 0, there is nothing stopping 3 wild card suit cards from suit 2 from being chosen as the 3 cards in the auto trident, need at least one card from suit 0 in order for it to count as a flush trident, i need to make a decision regarding whether the one mandatory suit card should be high or low , i think it should be high... just make it always the third one, the 1st card would always be the highest, then the 2nd second highest, 3rd 3rd highest obviously, so if we replace the 3rd one, the least high one with a card from the chosen suit that will be close to the target value This seems to work, Next i need to work on the flip and rotate, Done that , seems to work,
I’ve noticed the tridentHighlightPos doesn’t seem to work anymore
Only in some circumstances, filling trident hand space 21 or above causes the next trident to fail, it’s card’s are spread over 2 tridents
I was modding by triHandCardFilledArray
 Which is 26 instead of nonPrePostGameCardsEach
 Which is 24

If you cancel the trident build i don’t think any cards should be added to anything, i have made this so

I think i might actually be finished with the trident building stage now...?

Well, i wanted to add a button to allow the player to view the map, so as to consider the layout when building tridents, but perhaps i’ll come back to that, that is only for some game modes.
I need to make the pre and post game cards stand out more, i need to redesign the layout and colour scheme of cards, including something to indicate that a card is a flush card.
Should probably also have access to a help page when trident building