Now that i can place tridents i want to have that reflected over the network, the way i want to do this, is first of all i should emit the player’s whole trihand in order, to the other player, then i can simply emit the index number of their hand during play, will also need to emit the position so we know if it was rotated or flipped, this ‘position’ can be represented with one of 6 numbers
123
312
231
132
213
321

When the trident is point up rotations will appear to be going clockwise, when point down  anti clockwise, but actually it’s always just swapping left with right and then vert with left
I think the best way to notarise the position if with 3 numbers for rotation and a bool for flipped.

So

123	R0F0
312	R1F0
231	R2F0
132	R0F1
213	R1F1
321	R2F1

After a bit of thinking i’ve worked out that rotating will just increment the R number by one, use mod 3 so 3 becomes one, so rotating will just cycle between R0, R1, and R2, back to R0 etc, flipping will just toggle the F number either 0 or 1. And importantly rotation when already flipped will decrement the R number instead of increment. So you can rotate and flip as many times as you want, so long as i follow these rules i have a shorthand way on knowing the position

Then i just need the index number of the gameboard to know where it was placed

So i need to send the whole trihand to the either player, need to know
Playerindex
Value of each card in order, can derive suit from value, we know every 3rd card will be the end of a trident, the last 2 cards will be pre or post game cards if enabled
Need to know which suit is wildcard suit, if any
Rotation and flipped will always be 0 and false.

Sent that data across and displayed it under the player’s tri hand, it all looks ok. Now i need to emit the data for when it is placed on the gameboard-  but this can just be the index position of the trihand, index of gameboard and rotation and flip