Going to try to add the new deck building stage, make it so that after matchmaking we go to deck building, i should disable the trident building stages. I should alter the deal stage so that it doesn’t actually deal, but it does allow you to set the par handicap. The deck stage should use data in the PremadeDeck class to display all the possible tridents, the player will simply pick the correct number of tridents within their handicap, then we can continue to the game
Set up a variable called usePreMadeDeck

In the options stage
Created a new case in the gotostage method, created a new stage

I feel like i want to start over again, the code is getting messy with confusing animation stuff in the deal stage, confusing positions and updating stuff, using variables from different classes, using bloated button classes just to display a triangle , i’ll just press on

Set up the new stage, just the basics

Need to add the new stage to the touch logic and otherhit method of the tirbutton class
The deck building stage is now visible and the next stage button works on it
Now i need to display all possible premade tridents and allow the player to add and remove them to the trident hand. The handicap should be visible and a running total of the combined value of the tridents chosen should be displayed, you should not be able to add a trident to the trident hand if it would cause the total to exceed the handicap

The create buttons method in the deckBuildingStage need to set up 17 tridents in the available tridents array, need to set up 8 tridents in the trident hand array, - actually they must all be part of the same array, to be consistent with the way i’ve been doing it
Then i need to set the positions for all those. Then set the value of the available tridents based on the values of the tridents in the premadeDeck class, then when selecting the tridents rather than move them to the trident hand array, copy the values over to the trident hand array. Got the position working for the trident buttons that make up the trident hand, need to do the same for the available tridents then load in the values