I need to decide what the default handicap is, 8 tridents with 3 cards each is 24 cards, cards have value from 1 to 13, so average value is 7, so times 24 is 168. The highest value deck you could possibly pick with 8 trident would be 252, you would have 4 8s 4 9s 4 10s etc up to 4 13s, the lowest you could have is 84, although these number might not be attainable since each suit must emit one card. If i were to make that emitted card always be 7 then these numbers would always be attainable.
If you have a high handicap it means you are good, I could either subtract the handicap value from the player’s available total value, or i could work out the difference of the 2 player’s handicaps and subtract that instead. I think i’ll do the latter, as it makes more sense when players are setting their own handicap if one is over egging it the other player can do the same and it will even out

There should be a separate value to adjust , called total value, to adjust the total value that both players can use, so the average is 168 but if players agree to a higher total value it will result in a different game, forcing a lower value or higher value will make things interesting.

I’ll just have a series of buttons on the deck building screen, 2 to raise and lower your handicap,
2 to raise or lower the total value,

I don’t want to do anything complicated to police it, so you can just choose your own handicap, but the opponent will be able to see your handicap so you can adjust if you think they are doing too much. I should also display the win loss ratio of the player and a recommended handicap, maybe i can use a sort of ELO like chess? Elo starts at 1000, you divide the loser’s elo by the winners elo, times that by 100 and that is the elo change, you add the elo change to the winner and subtract it from the loser.

This is a pretty simple system which scales your increase in elo with the quality of opponent beaten. And inversely scales the decrease in elo with the quality of the opponent you lost to.

So display your elo, the opponent’s elo, it would be nice to translate elo into some kind of recommended handicap but i would need some data to do that properly. Display up and down arrows for raising and lowering handicap then more up and down arrows for raising and lowering total value

Handicap will be self policing, if the opponent lowers theirs too much, you can counter that by lowering yours as it will be the difference in the 2 values that matters.

Both players will be able to select a total value they think both players should be able to use (before handicap). Default is 0 which would be 168, can go up to +84 or down to -84, which would give 252 or 84 respectively.  The total value will be calculated by averaging the player’s specified total values and will then ensure that absolute value does not exceed the absolute value of the lowest specified absolute value, so if one player wants the total value to be 200 and one player wants it to be 180, the average will be 190, but that exceeds the lowest absolute value so set it to 180. If one wants 200 and one wants 100, set it to 150, as the absolute value, e.g. the difference from 168 is less than either proposed value... actually i guess it would be more fair to set the total value as just 168 if the players are trying to take it in different directions

Also need to make sure that the player’s value limit combined with handicap does not go under 84, as they would not be able to make a full hand

So display 2 buttons for value limit, display the value limit, 2 buttons for handicap, display the handicap, display the opponent’s handicap,
Display your elo and the opponent elo

Haven’t actually made the elo do anything yet, i need to save it to the players’ internal storage on their device

I need to emit the handicap and value limit, then on the server calculate what the value limit is for both players and what the handicap should be for each player,

Then the players should pick their cards up to the value limit. I could turn it around and rather than pick cards not exceeding the limit, instead say the player has a certain amount of points and you spend them to select cards, handicap gives you bonus points to spend, value limit can increase the points for both players